fixed 0 coverage error

+++ src/target_programs/program_11.py
@@ -1,6 +1,6 @@
-def program_11_sym(x: float):
-   z:float = 3**x
-   y:float = 0
+def program_11_sym(x: int):
+   z:int = 3**x
+   y:int = 0
    if z < 1:
       y = 0
    elif z < 2:
@@ -23,7 +23,7 @@ def program_11_sym(x: float):
 
 from target_programs.functions_to_approximate import neuzz_fn
 
-def program_11(x: float):
+def program_11(x: int):
    y:int = neuzz_fn(x)
    r:float = 0.0
    if y == 0:
